// src/app/models/advanced-search/search-types.enum.ts
export enum SearchTypesEnum {
  Text = 'text',
  TextPrimary = 'textPrimary',
  Number = 'number',
  Key = 'key',
  Date = 'date',
  Data = 'data',
  DataWithCheckbox = 'dataWithCheckbox',
  Boolean = 'boolean',
  Age = 'age',
  None = 'none',
  NumberIdentifier = 'numberIdentifier',
  StringIdentifier = 'stringIdentifier',
  In = 'in',
}